backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "fake-useragent>=2.2.0",
    "fastapi>=0.124.4",
    "google-generativeai>=0.8.5",
    "httpx>=0.28.1",
    "langchain-google-genai>=2.1.0",
    "loguru>=0.7.3",
    "openai>=2.12.0",
    "playwright>=1.49.0",
    "pydantic-settings>=2.12.0",
    "pydub>=0.25.1",
    "python-dotenv>=1.2.1",
    "tenacity>=9.0.0",
    "trafilatura>=2.0.0",
    "uvicorn[standard]>=0.38.0",
    # GCS Storage
    "google-cloud-storage>=2.18.0",
    # Phoenix LLMOps Tracing
    "arize-phoenix>=8.0.0",
    "openinference-instrumentation-langchain>=0.1.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.9",
]

[tool.uvicorn]
reload-exclude = [".venv"]